Added Holstein base distribution
A class for the Holstein base distribution has been writte, also added the setup.py in order to create a package to use easier for testing.
To correctly form the package also the varius __init__.py have been added.
Testing of the new class has been done and it works fine, only thing is that the probability of the diagrams drawn is often > 1 due to definition of it.
I have tested the generation and also if the Module can be run on GPU and works fine.