Adjusted MultiGaussian and added a first Conditioner
Finished the creation of the class for the multi variate Gaussian distribution adding also the TwoMoon one as a toy model target.
Then a MaskedLinear Module was created and tested for the generation of the MaskedConditioner one that will be used inside Normalizing Flows.
All has been tested so far to see if the things worked in terms of dimensionality of data input and output.

TODO: needs to implement better the mask generation inside the MaskedConditioner since I'm not counting that the parameters for a single z_i can be more than one, and so we need all of them to not depend on the variables after.