Finished Affine Masked Flow
The Masked Conditioner had a problem with the masked matrix, was upside down and I fixed it also inserting the fact that it should be different from input to hidden layers.
Created the Affine flow class in a simple way so that it can take a general conditioner and use it to perform forward and inverse.

All is fucking tested, so now the conditioner has the right properties and the affine flow works as wanted.