The model kind of works
I have added the Permutation layers, one that swaps variables randomly and one that flip the vector.
The model has been trained for the first time on the TwoMoon distribution using the reverse_kdl and it TRAINS, but is not able to reproduce the exact shape wanted.
Probably this is due to the fact that the masked conditioner I have inserted is highly limited, I will try to allow for larger layers to be created inside it.

I have also inserted inside the manager the equation to evaluate the log probability of the model in order to visualize the result of the training on the TwoMoon, and also the forward KDL loss function in order to try the train also with that.
Nevertheless in order to train on the forward KDL I need to be able to sample from TwoMoon and to do that I need to implement the rejection sampling method.