Restuctured the project
Created an Archetype folder in order to store all the Archetypes used inside the project.
Created a Transformer archetipe to define a general transformation with forward, inverse and log det function.
Rewrite the Affine flow in this architecture and also defined an Autoregressive class that couples the transformer to the Conditioner.
In this way I hope to generalize every transformer in order to make it autoregressive or coupling flow as one likes, only needs to use a different class to couple them.