Inserted the possibility of Multi Modal Gaussian
Inserted as a general distribution the multi modal gaussian and started the implementation of the diagrammatic part.
The idea is to create a base class that posses a static argument containing how the diagram is composed in order for all the transformation to know where to apply the right things.
For this reason I have created the DiaModule class, which might probabily better to not be a Module at all, probably is better only just a kind of container class.
Like DiaModifier, meaning that who inherits it is able to see and modify the diagram composition through set and get method.
Added also the first diagram permutation class in order to permute the time couples inside a specified block, it can also easily be done so that it permutes everithing inside every block.

I have also inserted a little bit more tests and documentation on old classes.