Created the Diagrammatic class
Changed the DiaModifier to Diagrammatic in order to define diagrammatic flows that can work on diagrams.
Diagrammatic now has the mean to store the diagram composition as it evolves in the flow and hand it over to the modules that needs it during the transformation.
It also gives the possibility to store changes in the composition for the transformation that modify the position of the variables inside the diagram, like permuations.
The Diagrammatic class and methods have been tested.

Also commented the diagrammatic permutation flows inserted previously that still needs to be tested.