Tested the Diagrammatic flows
Added some comments to the various scripts for readibility.
Added tests for the Diagrammatic swapping methods and modified a little the structure of Diagrammatic,
now to get the composition and modify it you need to be inheriting from it, but a print function has been
created to see the diagram composition on standard out put.