Added the DiaChecker
Added a new type of flow created in order to simply transform the blocks of the input vector in real diagrams by appling a series of transformations specific for every block  type.
Added the blocks type in order to define what are the different blocks.

The other things are small changes, like the test for the things I have added and a permutation that flips the couples inside a time ordered block.