Today was a mess
I started making changes in order to be able to construct the first network on the Holstein model but a lot of problems popped up.
Still I have inserted new permutators and chagned the name of existing Swap to Flip because I though it made more sense.
I also created a flow that add the creation time to the annhilation one but soon realized it's not good since doesn't bound from above the times and we need them to be less than tau.
Now I'm thinking on a solution using hyperbolic tangent to bound in the wanted range the times, but seems to have problems with the computation of the log determinant since the derivative is really close to zero.