Added a new transformer
Mainly started the design and construction of the main type of transformers that I will try to use in order to constrain the flow to wanted values.
In particular I'm talking of the Positive Affine that maps positive values in other positives.
There are other changes, like the OrderTime flow added, but they didn't worked due to a divergence problem in the loss that was given by the nearly zero derivative presente inside the sigmoid used in order to bound the times values.