Start creation of new approach
Complited the creation of Positive affine and COnstrained positive affine transformers for the new approach to the flow.
Created the first draft for the Block Coupling and Time Order Block Coupling flows that will allow to apply certain transformers to selected blocks without worring of the swapping problem.
So far they have been tested and are working with the wanted transformer, still I'm not too sure that they will work when the result will be back propagated, so you should try to backpropagate the log det obatined and see if errors occours.

I have also made slight changes to Diagrammatic just to make things easier to use inside other classes.