Finally it trains
I have added a lot of things in order to make it work, then tomorrow I will focus on make it clean.
I have changed CPAffine a lot in order to be able to insert another extra constrain that defines the possible maximum value of the variable we are using for numerical stability.
Also modified the formula to make it more numerically stable.
Created Coupling for both order and time of flight since I have figured out that constrains are also needed on them to train and the time of flight in particular also need to be grater than the higher phononic time.

So far things work but I need to clean up and write tests for everithing, but for now enjoy the fact that I was able to train a network for 10000 steps.
Probably adding more layer will kind of simplify things since no more too large values will be needed in the parameters and will be distributed along the various flows.