Started dev branch
Created this branch where I can write dirty code with a lot of if statement allowing me to see possible problems inside the chain of flows but make the code kind of unreadable.
In particular right now the code works since i have added a lot of debugging thanks to the several if and some little things to mantain numerical stability, like the 0.999999 instead of 1 in the generation of order and tm_fly.
Changed the way Baseholstein generates the elements since having a uniformly distributed order makes it harder to the flow to find the right result, make it exponential as well.
Both order and tm_fly are now distributed using a constrained exponential.